requirements.txt
Flask==2.3.3
Jinja2==3.1.2
psycopg2==2.9.7
Werkzeug==2.3.7
python-decouple==3.8
python-dotenv==1.0.0
